fix: Remove dependecies service from docker compose
This service is not needet to be built as a service with the rest of
them when running docker compose build/up. It is just needed for the
microservice containers to be built from. So the dependencies image is
now built separately and uploaded to docker hub and those other docker
files just pull from that image.